dandelionR: Single-cell immune repertoire trajectory analysis

dandelionR is an R package for performing single-cell immune repertoire trajectory analysis, based on the original python implementation. It provides the necessary functions to interface with scRepertoire and a custom implementation of an absorbing Markov chain for pseudotime inference, inspired by the Palantir Python package.

Main functions

setupVdjPseudobulk: Preprocess V(D)J Data for Pseudobulk Analysis.

vdjPseudobulk: Generate Pseudobulk V(D)J Feature Space.

markovProbability: Markov Chain Construction and Probability Calculation.

projectPseudotimeToCell: Project Pseudotime and Branch Probabilities to Single Cells.

Vignettes

See the package vignettes for detailed workflows: vignette('dandelionR')

Installation

To install from Bioconductor, use:

if (!requireNamespace('BiocManager', quietly = TRUE))
    install.packages('BiocManager')
BiocManager::install('dandelionR')

See also

Author

Maintainer: Kelvin Tuong z.tuong@uq.edu.au (ORCID)

Authors:

Jiawei Yu jiawei.yu@uq.edu.au (ORCID)

Nicholas Borcherding borcherding@wustl.edu (ORCID)