Example SCE Dataset that does not contain V(D)J information

The demo_sce object is a down-sampled demo dataset derived from Suo et al., 2024, Nature Biotechnology.
This dataset is used in vignettes to demonstrate workflows for V(D)J analysis.
For details, see the original publication at https://www.nature.com/articles/s41587-023-01734-7. The original Lymphoid cells data in h5ad format is available at https://developmental.cellatlas.io/fetal-immune.

Usage

data(demo_sce)

Format

A SingleCellExperiment object with the following slots:

colData

A minimall DataFrame containing metadata about each sample, corresponding to obs in AnnData (Python). The following columns are relevant for vignette usage:

anno_lvl_2_final_clean

Cell type annotations.

int_colData

A DataFrame containing additional assay metadata important for further analysis. Includes:

X_scvi: A dimensionality reduction matrix from the scVI model.

UMAP: A UMAP reduction matrix.

Source

Suo et al., 2024, Nature Biotechnology.
https://www.nature.com/articles/s41587-023-01734-7.

Examples

data(demo_sce)