markov_probability

Preprocessing data and Construct markov chain and calculate probabilities

markov_probability(
  milo,
  diffusionmap,
  diffustiontime = NULL,
  terminal_state,
  root_cell,
  pseudotime_key = "pseudotime",
  scale_components = TRUE,
  num_waypoints = 500
)

Arguments

milo

milo or SingelCellExperiment object, with pseudotime stored in colData, used to store the result and extract pseudotime. Pseudotime stored in milo has higher priority than the value provided through the diffusiontime parameter

diffusionmap

DiffusionMap object corresponds to milo

terminal_state

the index of the terminal state

root_cell

the index of the root state

pseudotime_key

the column name in the colData that holds the inferred pseudotime

scale_components

logical, If True, the components will be scale before constructing markov chain

num_waypoints

integer, 500L by default. Number of waypoints to sample to construct markov chain.

diffusiontime

if milo do not restore pseudotime, use this parameter to transfer it to function

Value

milo or SinglCellExperiment object with pseudotime, probabilities in its colData

Examples

# load data
data(pb.milo)
library(SingleCellExperiment)
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, saveRDS, setdiff, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
branch.tips <- c(540,54)
names(branch.tips) <- c("CD8+T", "CD4+T")
root <- 222
DPTroot <- paste0("DPT", root)
# extract pca
pca <- t(as.matrix(reducedDim(pb.milo, type = "PCA")))
# Run diffusion map on the PCA
library(destiny)
#> 
#> Attaching package: ‘destiny’
#> The following object is masked from ‘package:SummarizedExperiment’:
#> 
#>     distance
#> The following object is masked from ‘package:GenomicRanges’:
#> 
#>     distance
#> The following object is masked from ‘package:IRanges’:
#> 
#>     distance
dm <- DiffusionMap(t(pca),n_pcs=50, n_eigs = 10)
#> 'as(<dsCMatrix>, "dgTMatrix")' is deprecated.
#> Use 'as(as(., "generalMatrix"), "TsparseMatrix")' instead.
#> See help("Deprecated") and help("Matrix-deprecated").
dif.pse <- DPT(dm, tips = c(root, branch.tips), w_width = 0.1)
pb.milo$pseudotime <- dif.pse[[DPTroot]]

pb.milo <- markov_probability(
milo=pb.milo, 
diffusionmap=dm, 
terminal_state=branch.tips, 
root_cell=root, 
pseudotime_key="pseudotime")
#> Sampling and flocking waypoints...
#> Markov chain construction...
#> Loading required namespace: bluster
#> Computing fundamental matrix and absorption probabilities...
#> Project probabilites from waypoints to each pseudobulk...

#visualization of the results
plotPCA(pb.milo,  color_by = "CD8+T") + scale_color_gradientn(colors = pal)
#> Error: unable to find an inherited method for function ‘plotPCA’ for signature ‘object = "Milo"’
plotPCA(pb.milo,  color_by = "CD4+T") + scale_color_gradientn(colors = pal)
#> Error: unable to find an inherited method for function ‘plotPCA’ for signature ‘object = "Milo"’