project_pseudotime_to_cell

Function to project pseudotime & branch probabilities from pb.milo (pseudobulk) to milo (cell).

project_pseudotime_to_cell(milo, pb_milo, term_states, suffix = "")

Arguments

milo

SingleCellExperiment or milo object

pb_milo

pseudobulk data

term_states

vector of terminal states with branch_probabilities to be transferred

suffix

suffix to be added after the added column names, default ”

Value

subset of milo or SingleCellExperiment object where cell that do not belong to any neighbourhood are removed and projected pseudotime information stored colData

Examples

sce_vdj <- setup_vdj_pseudobulk(sce_vdj, 
                                already.productive = FALSE)
#> Checking productivity from productive_abT_VDJ, productive_abT_VJ ...
#> 23270 of cells filtered
#> checking allowed chain status...
#> 4 of cells filtered
#> Extract main TCR from v_call_abT_VDJ, d_call_abT_VDJ, j_call_abT_VDJ, v_call_abT_VJ, j_call_abT_VJ ...
#> Complete.
#> Filtering cells from v_call_abT_VDJ_main, j_call_abT_VDJ_main, v_call_abT_VJ_main, j_call_abT_VJ_main ...
#> 202 of cells filtered
#> 8524 of cells remain.
# Build Milo Object                                 
traj_milo <- miloR::Milo(sce_vdj)
milo_object <- miloR::buildGraph(traj_milo, k = 50, d = 20, reduced.dim = "X_scvi")
#> Constructing kNN graph with k:50
milo_object <- miloR::makeNhoods(milo_object, reduced_dims = "X_scvi", d = 20)
#> Checking valid object
#> Running refined sampling with reduced_dim

# Construct Pseudobulked VDJ Feature Space
pb.milo <- vdj_pseudobulk(milo_object, col_to_take = "anno_lvl_2_final_clean")
pbs = milo_object@nhoods
pb.milo <- scater::runPCA(pb.milo, assay.type = "Feature_space")

# Define root and branch tips
pca <- t(as.matrix(SingleCellExperiment::reducedDim(pb.milo, type = "PCA")))
branch.tips <- c(which.max(pca[2, ]), which.min(pca[2, ]))
names(branch.tips) <- c("CD8+T", "CD4+T")
root <- which.min(pca[1, ])

# Construct Diffusion Map
dm <- destiny::DiffusionMap(t(pca), n_pcs = 50, n_eigs = 10)
dif.pse <- destiny::DPT(dm, tips = c(root, branch.tips), w_width = 0.1)

#Markov Chain Construction 
pb.milo <- markov_probability(milo = pb.milo, 
                              diffusionmap = dm, 
                              terminal_state = branch.tips, 
                              diffusiontime = dif.pse[[paste0("DPT", root)]],
                              root_cell = root, 
                              pseudotime_key = "pseudotime")
#> Sampling and flocking waypoints...
#> Markov chain construction...
#> Computing fundamental matrix and absorption probabilities...
#> Project probabilites from waypoints to each pseudobulk...

# Project Pseudobulk Data
cdata <- project_pseudotime_to_cell(milo_object, pb.milo, branch.tips)
#> 22 number of cells removed due to not belonging to any neighbourhood
#> Error: subscript contains invalid names