Example Dataset for V(D)J Analysis

The sce_vdj object is a down-sampled demo dataset derived from Suo et al., 2024, Nature Biotechnology. This dataset is used in vignettes to demonstrate workflows for V(D)J analysis. For details, see the original publication at https://www.nature.com/articles/s41587-023-01734-7.

Usage

data(sce_vdj)

Format

A SingleCellExperiment object with the following slots:

colData

A DataFrame containing metadata about each sample, corresponding to obs in AnnData (Python). The following columns are relevant for vignette usage:

productive_(mode)_VDJ, productive_(mode)_VJ

Factors indicating whether the heavy or light chain is productive. mode refers to the extraction mode for V(D)J genes and can be one of:

'abT': TCRαβ

'gdT': TCRγδ

'B': BCR

Gene segment fields

Gene segment annotations with column names in the format (v/d/j)_call_(mode)_(VDJ/VJ). Examples include:

v_call_abT_VDJ: V gene for TCRαβ VDJ recombination

d_call_abT_VJ: D gene for TCRαβ VJ recombination

chain_status

A factor describing the receptor chain's status.

anno_lvl_2_final_clean

Cell type annotations.

int_colData

A DataFrame containing additional assay metadata important for further analysis. Includes:

X_scvi: A dimensionality reduction matrix from the scVI model.

UMAP: A UMAP reduction matrix.

Source

Suo et al., 2024, Nature Biotechnology. https://www.nature.com/articles/s41587-023-01734-7.

Examples

data(sce_vdj)