setup_vdj_pseudobulk

Function for data preprocessing. This function will filters the data based on productivity, chain status, subsets data, extracts main v(d)j, and removes the unmapping data

setup_vdj_pseudobulk(
  sce,
  mode_option = c("abT", "gdT", "B"),
  already.productive = TRUE,
  productive_cols = NULL,
  productive_vj = TRUE,
  productive_vdj = TRUE,
  allowed_chain_status = c("Single pair", "Extra pair", "Extra pair-exception",
    "Orphan VDJ", "Orphan VDJ-exception"),
  subsetby = NULL,
  groups = NULL,
  extract_cols = NULL,
  filter_pattern = ",|None|No_cotig",
  check_vj_mapping = c("v_call", "j_call"),
  check_vdj_mapping = c("v_call", "j_call"),
  check_extract_cols_mapping = NULL,
  remove_missing = TRUE
)

Arguments

sce

SingleCellExperiment object, vdj data should be contained in colData for filtering

mode_option

optional, mode for extraction the V(D)J genes.

already.productive

logical, whether the data is already filtered by productivity.

productive_cols

character vector, names of colData used in productivity filtering, with NULL by default

productive_vj

logical, with TRUE by default. Option in productivity filtering. If True, cell will only be kept if the main VJ chain is productive

productive_vdj

logical, with TRUE by default. Option in productivity filtering. If True, cell will only be kept if the main VDJ chain is productive

allowed_chain_status

character vectors, optional, if specified, the element should within c('single pair','Extra pair', 'Extra pair-exceptipn', 'Orphan VDJ','Orphan VDJ-exception')

subsetby

character, with NULL by default, name of one colData provided for sub-setting.

groups

character vector, with NULL by default, condition for sub-setting

extract_cols

character vector, with NULL by default. colData names where VDJ/VJ information is stored, these colData will be used instead of the standard colData

filter_pattern

character, optional, with ',|None|No_contig' by default.

check_vj_mapping

character vector, optional. elements should come from c('v_call', 'j_call'), with c('v_call', 'j_call') by default.

check_vdj_mapping

character vector, optional. elements should come from c('v_call','d_call', 'j_call'), with c('v_call', 'j_call') by default.

check_extract_cols_mapping

character vecter, with NULL by default

remove_missing

bool, True by default

Value

filtered SingleCellExperiment object

Details

data will undergo several process, including productivity filtering, chain status filtering, subseting, main TCR extracting, and unmapping data removing.

parameters for productivity filtering:

already.productive

If true, the function will skip the productivity filtering

If false, retain only productive entries in columns specified by mode_option or productive_cols

mode_option

function will check colData(s) named productive_{mode_option}_{type}, where type should be 'VDJ' or 'VJ' or both, depending on values of productive_vj and productive_vdj.

If set as 'NULL', the function needs the option 'extract_cols' to be specified

productive_cols

must be be specified when productivity filtering is need to conduct and mode_option is NULL.

where VDJ/VJ information is stored so that this will be used instead of the standard columns.

productive_vj, productive_vdj

If True, cell will only be kept if the main V(D)J chain is productive

parameter for chain status filtering: allowed_chain_status, chain status to be kept

parameters for subsetting: subsetby, groups

subsetting process will only be conducted when both parameters are provided. After subsetting, only the cell with {groups} feature in {subsetby} will be used for computing the VDJ feature space

parameter for main v(d)j extraction: extract_cols

parameters for unmapping filtering:

filter_pattern

pattern to be filtered form object.

If is set to be NULL, the unmaping filtering procees won't start

check_vj_mapping, check_vdj_mapping

only colData specified by these arguments (check_vj_mapping and check_vdj_mapping) will be checked for unclear mappings

check_extract_cols_mapping, related to extract_cols

Only colData specified by the argument will be checked for unclear mapping, the colData should first specified by extract_cols

remove_missing

If true, will remove cells with contigs matching the filter from the object.

If False, will mask them with a uniform value dependent on the column name.

Examples

# load data
data(exsce)
#> Warning: data set ‘exsce’ not found
# check the dimension
dim(exsce)
#> Error: object 'exsce' not found

# filtered the data
exsce<-setup_vdj_pseudobulk(sce = exsce,
                            mode_option = "abT",  # set the mode to αβTCR
                            already.productive = FALSE) # need to filter the unproductive cells
#> Error: object 'exsce' not found
# check the remaining dim
dim(exsce)
#> Error: object 'exsce' not found