Project Pseudotime and Branch Probabilities to Single Cells

This function projects pseudotime and branch probabilities from pseudobulk (pb.milo) data to single-cell resolution (milo). The results are stored in the colData of the milo object.

Usage

projectPseudotimeToCell(milo, pb_milo, term_states, suffix = "")

Arguments

milo

A SingleCellExperiment or Milo object. Represents single-cell data where pseudotime and branch probabilities will be projected.

pb_milo

A pseudobulk Milo object. Contains aggregated branch probabilities and pseudotime information to be transferred to single cells.

term_states

A named vector of terminal states, with branch probabilities to be transferred. The names should correspond to branches of interest.

suffix

Character. A suffix to be added to the new column names in colData. Default is an empty string ('').

Value

subset of milo or SingleCellExperiment object where cell that do not belong to any neighbourhood are removed and projected pseudotime information stored colData

Examples