demo data used in vignettes

sce_vdj - A down-sampled demo data from Suo et al 2024 Nature Biotechnology. See https://www.nature.com/articles/s41587-023-01734-7

Usage

data(sce_vdj)

Format

A SingleCellExperiment object with the following slots filled

colData

A DataFrame containing metadata about each sample, corresponding to obs in AnnData for python.
Below only describe the column used in the vignette:

productive_(mode)_VDJ, productive_(mode)_VJ

factor containing the information about whether the heavy/light chain is productive mode: mode for extraction of the V(D)J genes, could be 'abT'(TCRαβ), 'gdT'(TCRγδ) or 'B'(BCR)

Gene segment fields

with a pattern like (v/d/j)_call_(mode)_(VDJ/VJ)
For example:

v_call_abT_VDJ: V gene for TCRαβ VDJ recombination

d_call_abT_VJ: D gene for TCRαβ VJ recombination

chain_status

the status of the receptor chain

anno_lvl_2_final_clean

cell type annotation

int_colData

A DataFrame contains metadata of sample, mostly additional assay that is important for further analysis.
This int_colData only has the reduction matrix

X_scvi: result of scVI model reduction

UMAP: result of umap reduction

Source

https://www.nature.com/articles/s41587-023-01734-7

Examples

data(sce_vdj)